Moved exception throwing to Services. Random emojis route support a number to select specific amount of emojis

+++ src/main/java/rs/chat/service/EmojiService.java
@@ -24,7 +24,7 @@ public class EmojiService {
 		return this.emojiRepository.existsByName(name);
 	}
 
-	public List<Emoji> getRandomEmojis(int numberOfEmojis) {
+	public List<Emoji> getRandomEmojis(long numberOfEmojis) {
 		return this.emojiRepository.selectRandomEmojis(numberOfEmojis);
 	}
